requirements.txt
opencv-python-headless>=4
Pillow>=9
pytesseract>=0
numpy>=1
